+++ frontend/src/component/FindTab/ResultView.js
